+++ EmployeeGraphQL/Application/Services/Project/IProjectService.cs
@@ -5,7 +5,7 @@ public interface IProjectService
 {
     Task<PagedResult<ProjectResponse>> Projects(long departmentId, QueryOptions options);
     Task<Project> CreateProject(ProjectInput input, CancellationToken cancellationToken);
-    Task<Project> UpdateProject(int id, ProjectInput input, CancellationToken cancellationToken);
-    Task<bool> DeleteProject(int id, CancellationToken cancellationToken);
-    Task<Project> PublishProject(int id, CancellationToken cancellationToken);
+    Task<Project> UpdateProject(long id, ProjectInput input, CancellationToken cancellationToken);
+    Task<bool> DeleteProject(long id, CancellationToken cancellationToken);
+    Task<Project> PublishProject(long id, CancellationToken cancellationToken);
 }
